chore: split oversized modules for issue #61 without API regressions
