+++ src/studio_ui/components/library.py
@@ -2,1352 +2,16 @@
 
 from __future__ import annotations
 
-import base64
-import json
-from collections import OrderedDict
-from urllib.parse import quote, urlencode
+from fasthtml.common import H2, Div
 
-from fasthtml.common import H2, H3, A, Button, Details, Div, Form, Img, Input, Option, P, Script, Select, Span, Summary
-
-from studio_ui.common.title_utils import truncate_title
 from universal_iiif_core.library_catalog import ITEM_TYPES
-
-_STATE_STYLE = {
-    "saved": ("Da scaricare", "app-chip app-chip-neutral"),
-    "queued": ("In coda", "app-chip app-chip-accent"),
-    "downloading": ("Download", "app-chip app-chip-primary"),
-    "running": ("Download", "app-chip app-chip-primary"),
-    "partial": ("Parziale", "app-chip app-chip-warning"),
-    "complete": ("Completo", "app-chip app-chip-success"),
-    "error": ("Errore", "app-chip app-chip-danger"),
-}
-_CATEGORY_LABELS = {
-    "manoscritto": "Manoscritto",
-    "libro a stampa": "Libro a stampa",
-    "incunabolo": "Incunabolo",
-    "periodico": "Periodico",
-    "musica/spartito": "Musica/Spartito",
-    "mappa/atlante": "Mappa/Atlante",
-    "miscellanea": "Miscellanea",
-    "non classificato": "Non classificato",
-}
-_SORT_LABELS = {
-    "priority": "Priorita operativa",
-    "recent": "Aggiornati di recente",
-    "title_az": "Titolo A-Z",
-    "pages_desc": "Piu pagine",
-}
-_ACTION_BUTTON_CLS = {
-    "primary": "app-btn app-btn-primary",
-    "success": "app-btn app-btn-primary",
-    "accent": "app-btn app-btn-accent",
-    "danger": "app-btn app-btn-danger",
-    "warning": "app-btn app-btn-accent",
-    "neutral": "app-btn app-btn-neutral",
-    "info": "app-btn app-btn-info",
-    "auto": "app-btn app-btn-accent",
-}
-_LINK_BUTTON_CLS = {
-    "primary": "app-btn app-btn-primary",
-    "neutral": "app-btn app-btn-neutral",
-    "external": "app-btn app-btn-accent",
-    "muted": "app-btn app-btn-muted",
-}
-_LIBRARY_FILTER_KEYS = [
-    "q",
-    "state",
-    "library_filter",
-    "category",
-    "mode",
-    "view",
-    "action_required",
-    "sort_by",
-]
-_CATEGORY_SELECT_TONE = {
-    "manoscritto": (
-        "bg-indigo-50 border-indigo-300 text-indigo-700 dark:bg-indigo-500/20 "
-        "dark:border-indigo-500/45 dark:text-indigo-200"
-    ),
-    "libro a stampa": (
-        "bg-emerald-50 border-emerald-300 text-emerald-700 dark:bg-emerald-500/20 "
-        "dark:border-emerald-500/45 dark:text-emerald-200"
-    ),
-    "incunabolo": (
-        "bg-amber-50 border-amber-300 text-amber-700 dark:bg-amber-500/20 dark:border-amber-500/45 dark:text-amber-200"
-    ),
-    "periodico": ("bg-sky-50 border-sky-300 text-sky-700 dark:bg-sky-500/20 dark:border-sky-500/45 dark:text-sky-200"),
-    "musica/spartito": (
-        "bg-fuchsia-50 border-fuchsia-300 text-fuchsia-700 dark:bg-fuchsia-500/20 "
-        "dark:border-fuchsia-500/45 dark:text-fuchsia-200"
-    ),
-    "mappa/atlante": (
-        "bg-teal-50 border-teal-300 text-teal-700 dark:bg-teal-500/20 dark:border-teal-500/45 dark:text-teal-200"
-    ),
-    "miscellanea": (
-        "bg-violet-50 border-violet-300 text-violet-700 dark:bg-violet-500/20 "
-        "dark:border-violet-500/45 dark:text-violet-200"
-    ),
-    "non classificato": (
-        "bg-slate-100 border-slate-300 text-slate-700 dark:bg-slate-700/35 dark:border-slate-600 dark:text-slate-200"
-    ),
-}
-
-
-def _normalize_mode(mode: str | None, *, default_mode: str = "operativa") -> str:
-    fallback = "archivio" if str(default_mode or "").strip().lower() == "archivio" else "operativa"
-    value = str(mode or "").strip().lower()
-    return value if value in {"operativa", "archivio"} else fallback
-
-
-def _state_badge(state: str) -> Span:
-    label, cls = _STATE_STYLE.get(
-        (state or "saved").lower(),
-        ("Da scaricare", "app-chip app-chip-neutral"),
-    )
-    return Span(label, cls=cls)
-
-
-def _action_button(
-    label: str,
-    url: str,
-    tone: str = "neutral",
-    confirm: str | None = None,
-    hint: str | None = None,
-) -> Button:
-    kwargs = {
-        "cls": _ACTION_BUTTON_CLS[tone],
-        "hx_post": url,
-        "hx_target": "#library-page",
-        "hx_swap": "outerHTML show:none",
-        "hx_include": "#library-filters",
-    }
-    if confirm:
-        kwargs["hx_confirm"] = confirm
-    if hint:
-        kwargs["title"] = hint
-    return Button(label, **kwargs)
-
-
-def _link_button(label: str, href: str, tone: str = "neutral", *, external: bool = False):
-    if not href:
-        return Span(label, cls=_LINK_BUTTON_CLS["muted"])
-    kwargs = {"href": href, "cls": _LINK_BUTTON_CLS[tone]}
-    if external:
-        kwargs["target"] = "_blank"
-        kwargs["rel"] = "noreferrer"
-    return A(label, **kwargs)
-
-
-def _state_counts(docs: list[dict]) -> dict[str, int]:
-    out = {k: 0 for k in _STATE_STYLE}
-    for doc in docs:
-        key = str(doc.get("asset_state") or "saved").lower()
-        out[key] = out.get(key, 0) + 1
-    return out
-
-
-def _kpi_strip(docs: list[dict]) -> Div:
-    counts = _state_counts(docs)
-    kpis = [
-        ("Totale", len(docs), "text-slate-800 dark:text-slate-100"),
-        ("Completi", counts.get("complete", 0), "text-slate-700 dark:text-slate-200"),
-        ("Parziali", counts.get("partial", 0), "text-slate-700 dark:text-slate-200"),
-        (
-            "In coda",
-            counts.get("queued", 0) + counts.get("downloading", 0) + counts.get("running", 0),
-            "text-slate-700 dark:text-slate-200",
-        ),
-        ("Errori", counts.get("error", 0), "text-rose-600 dark:text-rose-300"),
-        ("Da scaricare", counts.get("saved", 0), "text-slate-600 dark:text-slate-300"),
-    ]
-    cards = [
-        Div(
-            P(label, cls="text-xs uppercase tracking-wide text-slate-500 dark:text-slate-400"),
-            P(str(value), cls=f"text-2xl font-bold {color}"),
-            cls=(
-                "rounded-xl border border-slate-200 dark:border-slate-700 "
-                "bg-white dark:bg-slate-800/60 p-3 min-w-[110px]"
-            ),
-        )
-        for label, value, color in kpis
-    ]
-    return Div(*cards, cls="grid grid-cols-2 md:grid-cols-3 xl:grid-cols-6 gap-3 mb-5")
-
-
-def _library_query_params(
-    *,
-    view: str,
-    q: str,
-    state: str,
-    library_filter: str,
-    category: str,
-    mode: str,
-    default_mode: str,
-    action_required: str,
-    sort_by: str,
-) -> dict[str, str]:
-    params: dict[str, str] = {}
-    if q:
-        params["q"] = q
-    if state:
-        params["state"] = state
-    if library_filter:
-        params["library_filter"] = library_filter
-    if category:
-        params["category"] = category
-    normalized_mode = _normalize_mode(mode, default_mode=default_mode)
-    normalized_default_mode = _normalize_mode(default_mode, default_mode="operativa")
-    if normalized_mode != normalized_default_mode:
-        params["mode"] = normalized_mode
-    if view and view != "grid":
-        params["view"] = view
-    if action_required and action_required != "0":
-        params["action_required"] = action_required
-    if sort_by:
-        params["sort_by"] = sort_by
-    return params
-
-
-def _library_url_with_filters(**kwargs) -> str:
-    query = urlencode(_library_query_params(**kwargs))
-    return f"/library?{query}" if query else "/library"
-
-
-def _render_mode_switch(
-    *,
-    view: str,
-    q: str,
-    state: str,
-    library_filter: str,
-    category: str,
-    mode: str,
-    default_mode: str,
-    action_required: str,
-    sort_by: str,
-) -> Div:
-    current_mode = _normalize_mode(mode, default_mode=default_mode)
-    normalized_default_mode = _normalize_mode(default_mode, default_mode="operativa")
-
-    operational_url = _library_url_with_filters(
-        view=view,
-        q=q,
-        state=state,
-        library_filter=library_filter,
-        category=category,
-        mode="operativa",
-        default_mode=normalized_default_mode,
-        action_required=action_required,
-        sort_by=sort_by,
-    )
-    archive_url = _library_url_with_filters(
-        view=view,
-        q=q,
-        state=state,
-        library_filter=library_filter,
-        category=category,
-        mode="archivio",
-        default_mode=normalized_default_mode,
-        action_required=action_required,
-        sort_by=sort_by,
-    )
-    operational_cls = (
-        "app-btn app-btn-accent font-semibold" if current_mode == "operativa" else "app-btn app-btn-neutral"
-    )
-    archive_cls = "app-btn app-btn-accent font-semibold" if current_mode == "archivio" else "app-btn app-btn-neutral"
-
-    return Div(
-        A(
-            "Vista Operativa",
-            href=operational_url,
-            hx_get=operational_url,
-            hx_target="#library-page",
-            hx_swap="outerHTML show:none",
-            hx_push_url="true",
-            cls=operational_cls,
-        ),
-        A(
-            "Vista Archivio",
-            href=archive_url,
-            hx_get=archive_url,
-            hx_target="#library-page",
-            hx_swap="outerHTML show:none",
-            hx_push_url="true",
-            cls=archive_cls,
-        ),
-        cls="flex flex-wrap items-center justify-end gap-2",
-    )
-
-
-def _render_filters(
-    *,
-    view: str,
-    q: str,
-    state: str,
-    library_filter: str,
-    category: str,
-    mode: str,
-    default_mode: str,
-    action_required: str,
-    sort_by: str,
-    libraries: list[str],
-    categories: list[str],
-) -> Details:
-    current_state = (state or "").strip().lower()
-    current_mode = _normalize_mode(mode, default_mode=default_mode)
-    current_action_required = (action_required or "0").strip()
-    normalized_default_mode = _normalize_mode(default_mode, default_mode="operativa")
-    default_sort = "title_az" if current_mode == "archivio" else "priority"
-    current_sort = (sort_by or "").strip() or default_sort
-    category_options = categories or list(ITEM_TYPES)
-    active_filters_count = 0
-    if (q or "").strip():
-        active_filters_count += 1
-    if current_state:
-        active_filters_count += 1
-    if (library_filter or "").strip():
-        active_filters_count += 1
-    if (category or "").strip():
-        active_filters_count += 1
-    if current_sort != default_sort:
-        active_filters_count += 1
-    if (view or "grid") != "grid":
-        active_filters_count += 1
-    if current_action_required == "1":
-        active_filters_count += 1
-    if current_mode != normalized_default_mode:
-        active_filters_count += 1
-
-    return Details(
-        Summary(
-            Div(
-                Span("Filtri", cls="text-sm font-semibold text-slate-800 dark:text-slate-100"),
-                Span(
-                    f"{active_filters_count} attivi" if active_filters_count > 0 else "Nessun filtro attivo",
-                    cls="text-xs text-slate-500 dark:text-slate-400",
-                ),
-                cls="flex items-center justify-between gap-2",
-            ),
-            cls="cursor-pointer list-none px-3 py-2",
-        ),
-        Form(
-            Div(
-                Input(
-                    type="text",
-                    name="q",
-                    value=q,
-                    placeholder="Cerca per titolo, segnatura, reference, ID o biblioteca",
-                    cls=(
-                        "px-3 py-2.5 bg-white dark:bg-slate-800 border border-slate-300 dark:border-slate-600 "
-                        "rounded text-slate-800 dark:text-slate-100 text-sm flex-1 min-w-[240px]"
-                    ),
-                ),
-                Button(
-                    "Filtra",
-                    cls="app-btn app-btn-primary",
-                    type="submit",
-                ),
-                A(
-                    "Reset",
-                    href="/library",
-                    hx_get="/library",
-                    hx_target="#library-page",
-                    hx_swap="outerHTML show:none",
-                    hx_push_url="true",
-                    id="library-reset-filters",
-                    cls="app-btn app-btn-neutral",
-                ),
-                cls="flex flex-wrap gap-2",
-            ),
-            Div(
-                Input(type="hidden", name="mode", value=current_mode),
-                Select(
-                    Option("Tutti gli stati", value=""),
-                    Option("Da scaricare", value="saved", selected=current_state == "saved"),
-                    Option("In coda", value="queued", selected=current_state == "queued"),
-                    Option("In download", value="downloading", selected=current_state == "downloading"),
-                    Option("Parziale", value="partial", selected=current_state == "partial"),
-                    Option("Completo", value="complete", selected=current_state == "complete"),
-                    Option("Errore", value="error", selected=current_state == "error"),
-                    name="state",
-                    cls=(
-                        "px-3 py-2.5 bg-white dark:bg-slate-800 border border-slate-300 dark:border-slate-600 "
-                        "rounded text-slate-800 dark:text-slate-100 text-sm"
-                    ),
-                ),
-                Select(
-                    Option("Tutte le biblioteche", value=""),
-                    *[Option(lib, value=lib, selected=library_filter == lib) for lib in libraries],
-                    name="library_filter",
-                    cls=(
-                        "px-3 py-2.5 bg-white dark:bg-slate-800 border border-slate-300 dark:border-slate-600 "
-                        "rounded text-slate-800 dark:text-slate-100 text-sm"
-                    ),
-                ),
-                Select(
-                    Option("Tutte le categorie", value=""),
-                    *[
-                        Option(_CATEGORY_LABELS.get(cat, cat.title()), value=cat, selected=category == cat)
-                        for cat in category_options
-                    ],
-                    name="category",
-                    cls=(
-                        "px-3 py-2.5 bg-white dark:bg-slate-800 border border-slate-300 dark:border-slate-600 "
-                        "rounded text-slate-800 dark:text-slate-100 text-sm"
-                    ),
-                ),
-                Select(
-                    *[Option(label, value=key, selected=current_sort == key) for key, label in _SORT_LABELS.items()],
-                    name="sort_by",
-                    cls=(
-                        "px-3 py-2.5 bg-white dark:bg-slate-800 border border-slate-300 dark:border-slate-600 "
-                        "rounded text-slate-800 dark:text-slate-100 text-sm"
-                    ),
-                ),
-                Select(
-                    Option("Grid", value="grid", selected=view == "grid"),
-                    Option("List", value="list", selected=view == "list"),
-                    name="view",
-                    cls=(
-                        "px-3 py-2.5 bg-white dark:bg-slate-800 border border-slate-300 dark:border-slate-600 "
-                        "rounded text-slate-800 dark:text-slate-100 text-sm"
-                    ),
-                ),
-                Select(
-                    Option("Tutti gli elementi", value="0", selected=current_action_required != "1"),
-                    Option("Solo elementi da gestire", value="1", selected=current_action_required == "1"),
-                    name="action_required",
-                    cls=(
-                        "px-3 py-2.5 bg-white dark:bg-slate-800 border border-slate-300 dark:border-slate-600 "
-                        "rounded text-slate-800 dark:text-slate-100 text-sm"
-                    ),
-                ),
-                cls="grid sm:grid-cols-2 xl:grid-cols-3 gap-2 mt-2",
-            ),
-            id="library-filters",
-            hx_get="/library",
-            hx_target="#library-page",
-            hx_swap="outerHTML show:none",
-            hx_push_url="true",
-            hx_trigger="submit, change delay:200ms from:select, keyup changed delay:400ms from:input[name='q']",
-            cls="px-3 pb-3",
-        ),
-        open=active_filters_count > 0,
-        id="library-filters-panel",
-        data_collapsible_key="filters",
-        cls=("rounded-xl border border-slate-200 dark:border-slate-700 bg-slate-50 dark:bg-slate-900/30 mb-5"),
-    )
-
-
-def _library_filters_persistence_script(default_mode: str = "operativa") -> Script:
-    safe_default_mode = "archivio" if str(default_mode or "").strip().lower() == "archivio" else "operativa"
-    return Script(
-        f"""
-        (function () {{
-            const STORAGE_KEY = 'ui.library.filters.v1';
-            const DETAILS_STORAGE_KEY = 'ui.library.collapsible.v1';
-            const DEFAULT_MODE = {json.dumps(safe_default_mode)};
-            const FILTER_KEYS = [
-                'q', 'state', 'library_filter', 'category', 'mode', 'view', 'action_required', 'sort_by'
-            ];
-
-            function defaultFilters() {{
-                return {{
-                    q: '',
-                    state: '',
-                    library_filter: '',
-                    category: '',
-                    mode: DEFAULT_MODE,
-                    view: 'grid',
-                    action_required: '0',
-                    sort_by: ''
-                }};
-            }}
-
-            function normalizeFilters(raw) {{
-                const base = defaultFilters();
-                const src = (raw && typeof raw === 'object') ? raw : {{}};
-                FILTER_KEYS.forEach((key) => {{
-                    base[key] = String(src[key] || '').trim();
-                }});
-                if (!base.mode) base.mode = DEFAULT_MODE;
-                if (!base.view) base.view = 'grid';
-                if (!base.action_required) base.action_required = '0';
-                return base;
-            }}
-
-            function readSavedFilters() {{
-                try {{
-                    const raw = localStorage.getItem(STORAGE_KEY);
-                    if (!raw) return null;
-                    return normalizeFilters(JSON.parse(raw));
-                }} catch (_e) {{
-                    return null;
-                }}
-            }}
-
-            function saveFilters(data) {{
-                try {{
-                    localStorage.setItem(STORAGE_KEY, JSON.stringify(normalizeFilters(data)));
-                }} catch (_e) {{
-                    /* ignore quota/storage errors */
-                }}
-            }}
-
-            function clearSavedFilters() {{
-                try {{
-                    localStorage.removeItem(STORAGE_KEY);
-                }} catch (_e) {{
-                    /* ignore */
-                }}
-            }}
-
-            function hasMeaningfulFilters(data) {{
-                const f = normalizeFilters(data);
-                return Boolean(
-                    f.q ||
-                    f.state ||
-                    f.library_filter ||
-                    f.category ||
-                    f.sort_by ||
-                    f.mode !== DEFAULT_MODE ||
-                    f.view !== 'grid' ||
-                    f.action_required !== '0'
-                );
-            }}
-
-            function collectFormFilters(form) {{
-                const payload = {{}};
-                FILTER_KEYS.forEach((key) => {{
-                    const input = form.querySelector('[name="' + key + '"]');
-                    payload[key] = input ? String(input.value || '').trim() : '';
-                }});
-                return normalizeFilters(payload);
-            }}
-
-            function setFormFilters(form, data) {{
-                const normalized = normalizeFilters(data);
-                FILTER_KEYS.forEach((key) => {{
-                    const input = form.querySelector('[name="' + key + '"]');
-                    if (!input) return;
-                    input.value = normalized[key];
-                }});
-            }}
-
-            function urlHasFilterParams() {{
-                try {{
-                    const params = new URLSearchParams(window.location.search || '');
-                    return FILTER_KEYS.some((key) => params.has(key));
-                }} catch (_e) {{
-                    return false;
-                }}
-            }}
-
-            function filtersToQuery(data) {{
-                const f = normalizeFilters(data);
-                const params = new URLSearchParams();
-                if (f.q) params.set('q', f.q);
-                if (f.state) params.set('state', f.state);
-                if (f.library_filter) params.set('library_filter', f.library_filter);
-                if (f.category) params.set('category', f.category);
-                if (f.sort_by) params.set('sort_by', f.sort_by);
-                if (f.mode !== DEFAULT_MODE) params.set('mode', f.mode);
-                if (f.view !== 'grid') params.set('view', f.view);
-                if (f.action_required !== '0') params.set('action_required', f.action_required);
-                return params.toString();
-            }}
-
-            function bindFormPersistence(form) {{
-                if (!form || form.dataset.persistBound === '1') return;
-                const persist = () => saveFilters(collectFormFilters(form));
-                form.addEventListener('change', persist);
-                form.addEventListener('submit', persist);
-                form.dataset.persistBound = '1';
-            }}
-
-            function bindResetAction() {{
-                const resetLink = document.getElementById('library-reset-filters');
-                if (!resetLink || resetLink.dataset.resetBound === '1') return;
-                resetLink.addEventListener('click', () => clearSavedFilters());
-                resetLink.dataset.resetBound = '1';
-            }}
-
-            function readCollapsibleState() {{
-                try {{
-                    const raw = localStorage.getItem(DETAILS_STORAGE_KEY);
-                    if (!raw) return {{}};
-                    const parsed = JSON.parse(raw);
-                    return parsed && typeof parsed === 'object' ? parsed : {{}};
-                }} catch (_e) {{
-                    return {{}};
-                }}
-            }}
-
-            function saveCollapsibleState(state) {{
-                try {{
-                    localStorage.setItem(DETAILS_STORAGE_KEY, JSON.stringify(state || {{}}));
-                }} catch (_e) {{
-                    /* ignore storage errors */
-                }}
-            }}
-
-            function bindCollapsiblePersistence() {{
-                const nodes = document.querySelectorAll('details[data-collapsible-key]');
-                if (!nodes || !nodes.length) return;
-                const state = readCollapsibleState();
-                nodes.forEach((el) => {{
-                    const key = String(el.getAttribute('data-collapsible-key') || '').trim();
-                    if (!key) return;
-                    if (Object.prototype.hasOwnProperty.call(state, key)) {{
-                        el.open = Boolean(state[key]);
-                    }}
-                    if (el.dataset.collapsibleBound === '1') return;
-                    el.addEventListener('toggle', () => {{
-                        const current = readCollapsibleState();
-                        current[key] = Boolean(el.open);
-                        saveCollapsibleState(current);
-                    }});
-                    el.dataset.collapsibleBound = '1';
-                }});
-            }}
-
-            function restoreFiltersIfNeeded(form) {{
-                if (urlHasFilterParams()) return false;
-                const saved = readSavedFilters();
-                if (!saved || !hasMeaningfulFilters(saved)) return false;
-                const query = filtersToQuery(saved);
-                if (!query) return false;
-
-                setFormFilters(form, saved);
-                saveFilters(saved);
-                const url = '/library?' + query;
-                try {{
-                    window.history.replaceState(window.history.state, '', url);
-                }} catch (_e) {{
-                    /* ignore */
-                }}
-
-                if (window.htmx && typeof window.htmx.ajax === 'function') {{
-                    window.htmx.ajax('GET', url, {{ target: '#library-page', swap: 'outerHTML show:none' }});
-                    return true;
-                }}
-                window.location.href = url;
-                return true;
-            }}
-
-            function initLibraryFiltersPersistence() {{
-                bindCollapsiblePersistence();
-                const form = document.getElementById('library-filters');
-                if (form) {{
-                    bindFormPersistence(form);
-                    bindResetAction();
-                    const restored = restoreFiltersIfNeeded(form);
-                    if (!restored) {{
-                        saveFilters(collectFormFilters(form));
-                    }}
-                }}
-            }}
-
-            if (!window.__libraryFiltersPersistenceBootstrapped) {{
-                window.__libraryFiltersPersistenceBootstrapped = true;
-                document.addEventListener('DOMContentLoaded', initLibraryFiltersPersistence);
-                document.body.addEventListener('htmx:afterSwap', (event) => {{
-                    const target = event && event.target;
-                    if (!target) return;
-                    if (target.id === 'app-main' || target.id === 'library-page') {{
-                        initLibraryFiltersPersistence();
-                    }}
-                }});
-            }}
-
-            initLibraryFiltersPersistence();
-        }})();
-        """
-    )
-
-
-def _category_select_cls(item_type: str) -> str:
-    tone = _CATEGORY_SELECT_TONE.get(item_type, _CATEGORY_SELECT_TONE["non classificato"])
-    return f"app-field min-w-[180px] px-2.5 py-1.5 text-sm font-medium {tone}"
-
-
-def _primary_action(doc: dict):
-    doc_id = quote(str(doc.get("id") or ""), safe="")
-    library = quote(str(doc.get("library") or "Unknown"), safe="")
-    state = str(doc.get("asset_state") or "saved").lower()
-    has_missing = bool(doc.get("has_missing_pages"))
-
-    if state == "partial" and has_missing:
-        return _action_button(
-            "🔁 Riprendi mancanti",
-            f"/api/library/retry_missing?doc_id={doc_id}&library={library}",
-            "accent",
-        )
-
-    if state in {"saved", "partial", "error"}:
-        return _action_button(
-            "⬇️ Scarica" if state == "saved" else "🔁 Riprova download",
-            f"/api/library/start_download?doc_id={doc_id}&library={library}",
-            "primary",
-        )
-
-    return None
-
-
-def _maintenance_actions(doc: dict) -> list[Button]:
-    doc_id = quote(str(doc.get("id") or ""), safe="")
-    library = quote(str(doc.get("library") or "Unknown"), safe="")
-    state = str(doc.get("asset_state") or "saved").lower()
-
-    actions = []
-
-    if state in {"partial", "error"}:
-        actions.append(
-            _action_button(
-                "🧹 Pulizia parziale",
-                f"/api/library/cleanup_partial?doc_id={doc_id}&library={library}",
-                "warning",
-            )
-        )
-    return actions
-
-
-def _delete_action(doc: dict) -> Button:
-    doc_id = quote(str(doc.get("id") or ""), safe="")
-    library = quote(str(doc.get("library") or "Unknown"), safe="")
-    return _action_button(
-        "🗑️ Elimina",
-        f"/api/library/delete?doc_id={doc_id}&library={library}",
-        "danger",
-        confirm="Confermi eliminazione completa del manoscritto locale?",
-    )
-
-
-def _category_form(doc: dict, item_type: str) -> Form:
-    doc_id = quote(str(doc.get("id") or ""), safe="")
-    library = quote(str(doc.get("library") or "Unknown"), safe="")
-    return Form(
-        Select(
-            *[
-                Option(_CATEGORY_LABELS.get(opt, opt.title()), value=opt, selected=item_type == opt)
-                for opt in ITEM_TYPES
-            ],
-            name="item_type",
-            onchange="this.form.requestSubmit()",
-            cls=_category_select_cls(item_type),
-        ),
-        hx_post=f"/api/library/set_type?doc_id={doc_id}&library={library}",
-        hx_target="#library-page",
-        hx_swap="outerHTML show:none",
-        hx_include="#library-filters",
-    )
-
-
-def _to_optional_bool(value) -> bool | None:
-    if value is None:
-        return None
-    if isinstance(value, bool):
-        return value
-    if isinstance(value, (int, float)):
-        return bool(value)
-    text = str(value).strip().lower()
-    if text in {"1", "true", "yes", "y"}:
-        return True
-    if text in {"0", "false", "no", "n", ""}:
-        return False
-    return None
-
-
-def _pdf_technical_info(doc: dict) -> tuple[str, str]:
-    source_raw = str(doc.get("pdf_source") or "").strip().lower()
-    if source_raw not in {"native", "images", "unknown"}:
-        native = _to_optional_bool(doc.get("has_native_pdf"))
-        if native is True:
-            source_raw = "native"
-        elif native is False:
-            source_raw = "images"
-        else:
-            source_raw = "unknown"
-
-    local_available = bool(_to_optional_bool(doc.get("pdf_local_available")))
-    local_count = int(doc.get("pdf_local_count") or 0)
-
-    source_map = {
-        "native": "nativa",
-        "images": "da immagini",
-        "unknown": "non nota",
-    }
-    source_label = source_map[source_raw]
-    local_count_label = str(max(local_count, 1) if local_available else 0)
-    return source_label, local_count_label
-
-
-def _compact_label(text: str) -> str:
-    return "".join(ch for ch in (text or "").lower().strip() if ch.isalnum())
-
-
-def library_card_dom_id(doc_id: str, library: str) -> str:
-    """Return a stable DOM id for one library card."""
-    raw = f"{library}::{doc_id}"
-    token = base64.urlsafe_b64encode(raw.encode("utf-8")).decode("ascii").rstrip("=")
-    return f"library-card-{token or 'item'}"
-
-
-def _doc_card_dom_id(doc: dict) -> str:
-    return library_card_dom_id(
-        str(doc.get("id") or ""),
-        str(doc.get("library") or "Unknown"),
-    )
-
-
-def _metadata_items(doc: dict) -> list[tuple[str, str]]:
-    raw = str(doc.get("metadata_json") or "").strip()
-    if not raw:
-        return []
-    try:
-        payload = json.loads(raw)
-    except Exception:
-        return []
-    if not isinstance(payload, dict):
-        return []
-    out: list[tuple[str, str]] = []
-    for key, value in payload.items():
-        label = str(key or "").strip()
-        text = str(value or "").strip()
-        if label and text:
-            out.append((label, text))
-    return out
-
-
-def _metadata_payload(doc: dict) -> str:
-    payload = {
-        "doc_id": str(doc.get("id") or ""),
-        "library": str(doc.get("library") or "Unknown"),
-        "card_id": _doc_card_dom_id(doc),
-        "title": str(doc.get("display_title") or doc.get("id") or "-"),
-        "shelfmark": str(doc.get("shelfmark") or doc.get("id") or "-"),
-        "date_label": str(doc.get("date_label") or ""),
-        "language_label": str(doc.get("language_label") or ""),
-        "reference_text": str(doc.get("reference_text") or ""),
-        "source_detail_url": str(doc.get("source_detail_url") or ""),
-        "metadata_items": _metadata_items(doc),
-    }
-    text = json.dumps(payload, ensure_ascii=True)
-    return base64.b64encode(text.encode("utf-8")).decode("ascii")
-
-
-def _tech_row(label: str, value: str) -> Div:
-    return Div(
-        Span(f"{label}:", cls="app-tech-key"),
-        Span(value, cls="app-tech-val"),
-        cls="app-tech-row",
-    )
-
-
-def _doc_card(doc: dict, *, compact: bool = False) -> Div:
-    state = str(doc.get("asset_state") or "saved").lower()
-    title = str(doc.get("display_title") or doc.get("id") or "-")
-    item_type = str(doc.get("item_type") or "non classificato")
-    total = int(doc.get("total_canvases") or 0)
-    downloaded = int(doc.get("downloaded_canvases") or 0)
-    progress = f"{downloaded}/{total}" if total > 0 else "0/0"
-    missing_count = len(doc.get("missing_pages") or [])
-    if missing_count <= 0 and total > downloaded:
-        missing_count = max(total - downloaded, 0)
-    detail_ref = str(doc.get("reference_text") or "")
-    source_detail_url = str(doc.get("source_detail_url") or "")
-    date_label = str(doc.get("date_label") or "")
-    lang = str(doc.get("language_label") or "")
-    shelfmark = str(doc.get("shelfmark") or doc.get("id") or "-")
-    library_name = str(doc.get("library") or "Unknown")
-    thumb_url = str(doc.get("thumbnail_url") or "")
-    pdf_source_label, pdf_local_count = _pdf_technical_info(doc)
-    compact_title = _compact_label(title)
-    compact_shelfmark = _compact_label(shelfmark)
-    compact_reference = _compact_label(detail_ref)
-    show_reference = bool(detail_ref) and compact_reference not in {compact_title, compact_shelfmark}
-
-    doc_id = quote(str(doc.get("id") or ""), safe="")
-    library = quote(str(doc.get("library") or "Unknown"), safe="")
-    studio_href = f"/studio?doc_id={doc_id}&library={library}"
-
-    thumbnail_block = (
-        Img(
-            src=thumb_url,
-            cls=(
-                "w-full md:w-32 h-44 md:h-40 object-cover rounded-lg border border-slate-200 dark:border-slate-700 "
-                "bg-slate-100 dark:bg-slate-800"
-            ),
-        )
-        if thumb_url
-        else Div(
-            "Anteprima non disponibile",
-            cls=(
-                "w-full md:w-32 h-44 md:h-40 rounded-lg border border-dashed border-slate-300 dark:border-slate-700 "
-                "text-xs text-slate-500 dark:text-slate-400 flex items-center justify-center text-center px-2"
-            ),
-        )
-    )
-
-    utility_actions = [
-        _link_button("📖 Apri Studio", studio_href, tone="primary"),
-        _link_button("🔗 Scheda catalogo ↗", source_detail_url, tone="external", external=True),
-        Button(
-            "🧾 Metadati",
-            type="button",
-            cls=_ACTION_BUTTON_CLS["info"],
-            data_payload=_metadata_payload(doc),
-            onclick="openLibraryMetadata(this.dataset.payload)",
-        ),
-    ]
-    primary_action = _primary_action(doc)
-    action_buttons = [*utility_actions]
-    if primary_action is not None:
-        action_buttons.append(primary_action)
-    action_buttons.extend(_maintenance_actions(doc))
-
-    media_badges = Div(
-        _state_badge(state),
-        cls="flex flex-wrap gap-2 mt-2",
-    )
-    technical_rows = Div(
-        _tech_row("Pagine scaricate", progress),
-        _tech_row("Pagine mancanti", str(missing_count)),
-        _tech_row("Data", date_label or "-"),
-        _tech_row("Lingua", lang or "-"),
-        _tech_row("Sorgente PDF", pdf_source_label),
-        _tech_row("PDF locali", pdf_local_count),
-        cls="app-tech-list mt-2",
-    )
-    media_column = Div(
-        thumbnail_block,
-        media_badges,
-        technical_rows,
-        cls="w-full md:w-44 shrink-0",
-    )
-
-    card_title = truncate_title(title, max_len=70, suffix="[...]")
-    headline = Div(
-        H3(
-            card_title,
-            title=title,
-            cls="text-base md:text-lg font-bold text-slate-900 dark:text-slate-100 leading-tight",
-        ),
-        Div(
-            Span(library_name, cls="text-sm text-slate-700 dark:text-slate-200 font-semibold"),
-            _category_form(doc, item_type),
-            cls="flex flex-wrap items-center gap-2",
-        ),
-        P(f"Segnatura: {shelfmark}", cls="text-sm text-slate-500 dark:text-slate-400"),
-        cls="space-y-1 min-w-0",
-    )
-
-    return Div(
-        media_column,
-        Div(
-            Div(
-                headline,
-                cls="flex items-start justify-between gap-3",
-            ),
-            (
-                P(detail_ref, cls="text-sm text-slate-600 dark:text-slate-300 border-l-2 border-slate-300 pl-2")
-                if show_reference
-                else Div()
-            ),
-            Div(
-                *action_buttons,
-                _delete_action(doc),
-                cls="mt-auto flex flex-wrap items-center gap-2 pt-1",
-            ),
-            cls="space-y-3 flex-1 min-w-0 flex flex-col",
-        ),
-        cls=(
-            "flex flex-col md:flex-row gap-4 rounded-xl border border-slate-200 dark:border-slate-700 "
-            + ("bg-slate-50 dark:bg-slate-800/55 p-3" if compact else "bg-slate-50 dark:bg-slate-800/75 p-4")
-        ),
-        id=_doc_card_dom_id(doc),
-        data_library_card="1",
-    )
 
 
 def render_library_card(doc: dict, *, compact: bool = False) -> Div:
     """Render a single library card for targeted HTMX swaps."""
-    return _doc_card(doc, compact=compact)
+    from .library_cards import render_library_card as _render_library_card
 
-
-def _render_operational_list(docs: list[dict], view: str) -> Div:
-    buckets = OrderedDict(
-        [
-            ("critici", {"label": "Critici", "entries": []}),
-            ("in_corso", {"label": "In corso", "entries": []}),
-            ("da_scaricare", {"label": "Da scaricare", "entries": []}),
-            ("completati", {"label": "Completati", "entries": []}),
-        ]
-    )
-
-    for doc in docs:
-        state = str(doc.get("asset_state") or "saved").lower()
-        if state == "error" or (state == "partial" and bool(doc.get("has_missing_pages"))):
-            buckets["critici"]["entries"].append(doc)
-        elif state in {"downloading", "running", "queued"}:
-            buckets["in_corso"]["entries"].append(doc)
-        elif state in {"saved", "partial"}:
-            buckets["da_scaricare"]["entries"].append(doc)
-        else:
-            buckets["completati"]["entries"].append(doc)
-
-    sections = []
-    grid_cls = "space-y-3" if view == "list" else "grid xl:grid-cols-2 gap-3"
-    for data in buckets.values():
-        entries = data["entries"]
-        if not entries:
-            continue
-        cards = [_doc_card(doc, compact=view == "list") for doc in entries]
-        sections.append(
-            Div(
-                Div(
-                    H3(data["label"], cls="text-base font-semibold text-slate-800 dark:text-slate-200"),
-                    Span(str(len(entries)), cls="text-sm text-slate-500 dark:text-slate-400"),
-                    cls="flex items-center justify-between mb-2",
-                ),
-                Div(*cards, cls=grid_cls),
-                cls="mb-5",
-            )
-        )
-    return Div(*sections, id="library-content")
-
-
-def _render_archive_list(docs: list[dict], view: str) -> Div:
-    grouped: OrderedDict[str, OrderedDict[str, list[dict]]] = OrderedDict()
-    for doc in docs:
-        lib = str(doc.get("library") or "Unknown")
-        cat = str(doc.get("item_type") or "non classificato")
-        grouped.setdefault(lib, OrderedDict()).setdefault(cat, []).append(doc)
-
-    sections = []
-    for lib, by_category in grouped.items():
-        lib_key = base64.urlsafe_b64encode(lib.encode("utf-8")).decode("ascii").rstrip("=") or "library"
-        category_blocks = []
-        for cat in ITEM_TYPES:
-            entries = by_category.get(cat)
-            if not entries:
-                continue
-            cards = [_doc_card(doc, compact=view == "list") for doc in entries]
-            grid_cls = "space-y-3" if view == "list" else "grid xl:grid-cols-2 gap-3"
-            category_blocks.append(
-                Div(
-                    Div(
-                        H3(
-                            _CATEGORY_LABELS.get(cat, cat.title()),
-                            cls="text-base font-semibold text-slate-700 dark:text-slate-200",
-                        ),
-                        Span(str(len(entries)), cls="text-sm text-slate-500 dark:text-slate-400"),
-                        cls="flex items-center justify-between mb-2",
-                    ),
-                    Div(*cards, cls=grid_cls),
-                    cls="mb-4",
-                )
-            )
-
-        sections.append(
-            Details(
-                Summary(
-                    Div(
-                        H3(lib, cls="text-lg font-bold text-slate-800 dark:text-slate-100"),
-                        Span(
-                            f"{sum(len(v) for v in by_category.values())} elementi",
-                            cls="text-sm text-slate-500 dark:text-slate-400",
-                        ),
-                        cls="flex items-center justify-between",
-                    ),
-                    cls="cursor-pointer list-none",
-                ),
-                Div(*category_blocks, cls="mt-3"),
-                open=True,
-                data_collapsible_key=f"archive:{lib_key}",
-                cls="rounded-xl border border-slate-200 dark:border-slate-700 bg-white dark:bg-slate-900/20 p-3 mb-4",
-            )
-        )
-
-    return Div(*sections, id="library-content")
-
-
-def render_library_list(docs: list[dict], view: str = "grid", mode: str = "operativa") -> Div:
-    """Render grouped library entries in grid/list mode."""
-    if not docs:
-        return Div(
-            P("Nessun documento trovato con i filtri correnti.", cls="text-sm text-slate-500 dark:text-slate-400"),
-            id="library-content",
-        )
-
-    if (mode or "operativa").lower() == "archivio":
-        return _render_archive_list(docs, view)
-    return _render_operational_list(docs, view)
-
-
-def _metadata_drawer() -> Div:
-    return Div(
-        Div(
-            id="library-meta-overlay",
-            onclick="closeLibraryMetadata(event)",
-            cls="hidden fixed inset-0 z-40 bg-slate-900/50",
-        ),
-        Div(
-            Div(
-                Div(
-                    H3(
-                        "Metadati documento",
-                        id="library-meta-title",
-                        cls="text-lg font-semibold text-slate-900 dark:text-slate-100",
-                    ),
-                    Button(
-                        "✕",
-                        type="button",
-                        onclick="closeLibraryMetadata()",
-                        cls="text-lg text-slate-500 hover:text-slate-700 dark:text-slate-300 dark:hover:text-slate-100",
-                    ),
-                    cls="flex items-center justify-between border-b border-slate-200 dark:border-slate-700 px-4 py-3",
-                ),
-                Div(
-                    Div(
-                        Div(
-                            Span("Segnatura:", cls="app-tech-key"),
-                            Span("-", id="library-meta-shelfmark", cls="app-tech-val"),
-                            cls="app-tech-row",
-                        ),
-                        Div(
-                            Span("Data:", cls="app-tech-key"),
-                            Span("-", id="library-meta-date", cls="app-tech-val"),
-                            cls="app-tech-row",
-                        ),
-                        Div(
-                            Span("Lingua:", cls="app-tech-key"),
-                            Span("-", id="library-meta-language", cls="app-tech-val"),
-                            cls="app-tech-row",
-                        ),
-                        id="library-meta-tech",
-                        cls="app-tech-list",
-                    ),
-                    P("", id="library-meta-reference", cls="text-sm text-slate-600 dark:text-slate-300"),
-                    Div(id="library-meta-items", cls="space-y-1"),
-                    cls="p-4 space-y-3 overflow-y-auto flex-1",
-                ),
-                Div(
-                    Button(
-                        "🧾 Aggiorna metadati",
-                        id="library-meta-refresh",
-                        type="button",
-                        cls=_ACTION_BUTTON_CLS["info"],
-                        hx_target="#library-page",
-                        hx_swap="outerHTML show:none",
-                        hx_include="#library-filters",
-                    ),
-                    A(
-                        "🔗 Scheda catalogo ↗",
-                        id="library-meta-catalog",
-                        href="#",
-                        target="_blank",
-                        rel="noreferrer",
-                        cls=_LINK_BUTTON_CLS["external"],
-                    ),
-                    cls="border-t border-slate-200 dark:border-slate-700 px-4 py-3 flex flex-wrap gap-2",
-                ),
-                cls=(
-                    "flex h-full flex-col bg-white dark:bg-slate-900 shadow-2xl "
-                    "rounded-t-2xl sm:rounded-none sm:rounded-l-2xl"
-                ),
-            ),
-            id="library-meta-sheet",
-            cls=(
-                "hidden fixed z-50 inset-x-0 bottom-0 max-h-[85vh] "
-                "sm:inset-y-0 sm:right-0 sm:left-auto sm:h-full sm:max-h-none sm:w-[560px]"
-            ),
-        ),
-        Script(
-            """(function () {
-                function isLibraryMutation(detail) {
-                    const path = (detail && detail.requestConfig && detail.requestConfig.path) || '';
-                    if (typeof path === 'string' && path.startsWith('/api/library/')) return true;
-                    const responseUrl = (detail && detail.xhr && detail.xhr.responseURL) || '';
-                    return responseUrl.indexOf('/api/library/') !== -1;
-                }
-
-                function isMetadataRefreshRequest(detail) {
-                    const path = String((detail && detail.requestConfig && detail.requestConfig.path) || '');
-                    if (path.indexOf('/api/library/refresh_metadata') !== -1) return true;
-                    const trigger = detail && detail.elt;
-                    if (trigger && trigger.id === 'library-meta-refresh') return true;
-                    return false;
-                }
-
-                function renderMetaRows(items) {
-                    const holder = document.getElementById('library-meta-items');
-                    if (!holder) return;
-                    holder.innerHTML = '';
-                    if (!Array.isArray(items) || !items.length) {
-                        const empty = document.createElement('p');
-                        empty.className = 'text-sm text-slate-500 dark:text-slate-400';
-                        empty.textContent = 'Nessun metadato dettagliato disponibile.';
-                        holder.appendChild(empty);
-                        return;
-                    }
-                    items.forEach((entry) => {
-                        if (!Array.isArray(entry) || entry.length < 2) return;
-                        const row = document.createElement('div');
-                        row.className = 'leading-snug';
-                        const k = document.createElement('span');
-                        k.className = 'text-sm text-slate-500 dark:text-slate-400';
-                        k.textContent = String(entry[0] || '') + ': ';
-                        const v = document.createElement('span');
-                        v.className = 'text-sm text-slate-700 dark:text-slate-200';
-                        v.textContent = String(entry[1] || '');
-                        row.appendChild(k);
-                        row.appendChild(v);
-                        holder.appendChild(row);
-                    });
-                }
-
-                function parsePayload(encoded) {
-                    if (!encoded) return null;
-                    try {
-                        return JSON.parse(atob(encoded));
-                    } catch (_e) {
-                        return null;
-                    }
-                }
-
-                window.openLibraryMetadata = function (encoded) {
-                    const data = parsePayload(encoded);
-                    if (!data) return;
-
-                    const overlay = document.getElementById('library-meta-overlay');
-                    const sheet = document.getElementById('library-meta-sheet');
-                    const title = document.getElementById('library-meta-title');
-                    const shelfmark = document.getElementById('library-meta-shelfmark');
-                    const dateLabel = document.getElementById('library-meta-date');
-                    const language = document.getElementById('library-meta-language');
-                    const reference = document.getElementById('library-meta-reference');
-                    const refresh = document.getElementById('library-meta-refresh');
-                    const catalog = document.getElementById('library-meta-catalog');
-                    if (
-                        !overlay || !sheet || !title || !shelfmark || !dateLabel || !language ||
-                        !reference || !refresh || !catalog
-                    ) return;
-
-                    title.textContent = data.title || 'Metadati documento';
-                    shelfmark.textContent = data.shelfmark || '-';
-                    dateLabel.textContent = data.date_label || '-';
-                    language.textContent = data.language_label || '-';
-
-                    reference.textContent = data.reference_text || '';
-                    reference.classList.toggle('hidden', !data.reference_text);
-                    renderMetaRows(data.metadata_items || []);
-
-                    const refreshUrl = '/api/library/refresh_metadata?doc_id='
-                        + encodeURIComponent(String(data.doc_id || ''))
-                        + '&library='
-                        + encodeURIComponent(String(data.library || 'Unknown'));
-                    const targetCard = String(data.card_id || '').trim();
-                    const cardRefreshUrl = refreshUrl + (targetCard ? '&card_only=1' : '');
-                    refresh.setAttribute('data-card-id', targetCard);
-                    refresh.setAttribute('hx-target', targetCard ? ('#' + targetCard) : '#library-page');
-                    refresh.setAttribute('hx-swap', 'outerHTML show:none');
-                    refresh.setAttribute('hx-post', cardRefreshUrl);
-                    if (window.htmx && typeof window.htmx.process === 'function') {
-                        window.htmx.process(refresh);
-                    }
-
-                    if (data.source_detail_url) {
-                        catalog.setAttribute('href', data.source_detail_url);
-                        catalog.classList.remove('hidden');
-                    } else {
-                        catalog.setAttribute('href', '#');
-                        catalog.classList.add('hidden');
-                    }
-
-                    overlay.classList.remove('hidden');
-                    sheet.classList.remove('hidden');
-                    document.body.classList.add('overflow-hidden');
-                };
-
-                window.closeLibraryMetadata = function (event) {
-                    if (event && event.target && event.target.id !== 'library-meta-overlay') return;
-                    const overlay = document.getElementById('library-meta-overlay');
-                    const sheet = document.getElementById('library-meta-sheet');
-                    if (overlay) overlay.classList.add('hidden');
-                    if (sheet) sheet.classList.add('hidden');
-                    document.body.classList.remove('overflow-hidden');
-                };
-
-                document.addEventListener('keydown', (event) => {
-                    if (event.key === 'Escape') window.closeLibraryMetadata();
-                });
-
-                if (!window.__libraryScrollRestoreBound) {
-                    window.__libraryScrollRestoreBound = true;
-
-                    document.body.addEventListener('htmx:beforeRequest', (event) => {
-                        if (!isLibraryMutation(event.detail)) return;
-                        const isMetaRefresh = isMetadataRefreshRequest(event.detail);
-                        const appMain = document.getElementById('app-main');
-                        if (appMain) {
-                            window.__libraryScrollTop = appMain.scrollTop || 0;
-                        }
-                        if (isMetaRefresh) {
-                            const refreshBtn = document.getElementById('library-meta-refresh');
-                            const explicitCardId = refreshBtn
-                                ? String(refreshBtn.getAttribute('data-card-id') || '').trim()
-                                : '';
-                            const targetSelector = refreshBtn ? String(refreshBtn.getAttribute('hx-target') || '') : '';
-                            const targetCardId = explicitCardId || (
-                                targetSelector.startsWith('#library-card-')
-                                    ? targetSelector.slice(1)
-                                    : ''
-                            );
-                            window.__libraryMetaRefreshContext = {
-                                cardId: targetCardId
-                            };
-                        } else if (typeof window.closeLibraryMetadata === 'function') {
-                            window.closeLibraryMetadata();
-                            window.__libraryMetaRefreshContext = null;
-                        }
-                        if (!isMetaRefresh) {
-                            const active = document.activeElement;
-                            if (active && typeof active.blur === 'function') active.blur();
-                        }
-                    });
-
-                    document.body.addEventListener('htmx:afterSwap', (event) => {
-                        if (!isLibraryMutation(event.detail)) return;
-                        if (!event.detail || !event.detail.target) return;
-                        const targetId = event.detail.target.id || '';
-
-                        if (isMetadataRefreshRequest(event.detail)) {
-                            const ctx = window.__libraryMetaRefreshContext || {};
-                            const cardId = String(ctx.cardId || '').trim();
-                            if (cardId && typeof window.openLibraryMetadata === 'function') {
-                                const card = document.getElementById(cardId);
-                                const opener = card ? card.querySelector('[data-payload]') : null;
-                                const payload = opener && opener.dataset ? opener.dataset.payload : '';
-                                if (payload) {
-                                    window.openLibraryMetadata(payload);
-                                }
-                            }
-                            window.__libraryMetaRefreshContext = null;
-                        }
-
-                        if (targetId !== 'app-main' && targetId !== 'library-page') return;
-                        const appMain = document.getElementById('app-main');
-                        if (!appMain) return;
-                        const targetTop = Number(window.__libraryScrollTop);
-                        if (!Number.isFinite(targetTop)) return;
-                        window.requestAnimationFrame(() => {
-                            appMain.scrollTop = Math.max(0, targetTop);
-                        });
-                    });
-                }
-            })();"""
-        ),
-    )
+    return _render_library_card(doc, compact=compact)
 
 
 def render_library_page(
@@ -1366,6 +30,14 @@ def render_library_page(
     categories: list[str] | None = None,
 ) -> Div:
     """Render the full Local Library page."""
+    from .library_cards import _kpi_strip, _metadata_drawer, render_library_list
+    from .library_filters import (
+        _library_filters_persistence_script,
+        _normalize_mode,
+        _render_filters,
+        _render_mode_switch,
+    )
+
     libraries = libraries or []
     categories = categories or list(ITEM_TYPES)
     normalized_default_mode = _normalize_mode(default_mode, default_mode="operativa")
@@ -1407,3 +79,6 @@ def render_library_page(
         cls="p-6 max-w-7xl mx-auto",
         id="library-page",
     )
+
+
+__all__ = ["render_library_card", "render_library_page"]
