+++ src/universal_iiif_core/logic/downloader.py
@@ -1,11 +1,8 @@
 from __future__ import annotations
 
-import json
-import shutil
 import time
 from collections import deque
 from collections.abc import Callable
-from concurrent.futures import ThreadPoolExecutor, as_completed
 from contextlib import suppress
 from pathlib import Path
 from secrets import SystemRandom
@@ -13,17 +10,14 @@ from typing import Any
 
 import requests
 from PIL import Image
-from requests import RequestException
-from tqdm import tqdm
 
 from ..config_manager import get_config_manager
-from ..export_studio import build_professional_pdf
 from ..iiif_tiles import stitch_iiif_tiles_to_jpeg
 from ..library_catalog import parse_manifest_catalog
 from ..logger import get_download_logger
-from ..pdf_utils import convert_pdf_to_images
+from ..pdf_utils import convert_pdf_to_images  # noqa: F401 - preserved for monkeypatch compatibility in tests
 from ..services.storage.vault_manager import VaultManager
-from ..utils import DEFAULT_HEADERS, clean_dir, ensure_dir, get_json, save_json
+from ..utils import DEFAULT_HEADERS, ensure_dir, get_json, save_json
 from .download_helpers import derive_identifier
 
 # Constants
@@ -611,456 +605,11 @@ class IIIFDownloader:
             return thumbnail.get("@id") or thumbnail.get("id")
         return thumbnail
 
-    def create_pdf(self, files=None):
-        """Generates a professional PDF with cover and colophon."""
-        if files is None:
-            files = sorted(
-                [str(p) for p in self.scans_dir.iterdir() if p.name.startswith("pag_") and p.suffix.lower() == ".jpg"]
-            )
-        if not files:
-            return
 
-        output_path = self._determine_pdf_output_path()
-        cm = get_config_manager()
-        cover_cfg = cm.get_setting("pdf.cover", {})
-        logo_bytes = self._load_logo_bytes(cover_cfg.get("logo_path", ""))
-        curator = cover_cfg.get("curator", "")
-        desc = cover_cfg.get("description", "")
-        transcription_json = None
-        if self.ocr_path.exists():
-            transcription_json = get_json(str(self.ocr_path))
+from .downloader_pdf import attach_pdf_methods  # noqa: E402
+from .downloader_runtime import attach_runtime_methods  # noqa: E402
 
-        try:
-            self.logger.info(f"Generating PDF for {len(files)} pages...")
-            self.logger.info("Creating Professional PDF with Cover & Colophon...")
+attach_pdf_methods(IIIFDownloader)
+attach_runtime_methods(IIIFDownloader)
 
-            with tqdm(total=len(files), desc="Generating PDF", unit="page") as pbar:
-
-                def update_progress(current, total):
-                    pbar.update(1)
-
-                build_professional_pdf(
-                    doc_dir=self.doc_dir,
-                    output_path=output_path,
-                    selected_pages=list(range(1, len(files) + 1)),
-                    cover_title=str(self.label),
-                    cover_curator=curator,
-                    cover_description=desc or str(self.manifest.get("description", "")),
-                    manifest_meta=self.manifest,
-                    transcription_json=transcription_json,
-                    mode="Solo immagini",
-                    compression="Standard",
-                    source_url=self.manifest_url,
-                    cover_logo_bytes=logo_bytes,
-                    progress_callback=update_progress,
-                )
-
-            self.logger.info(f"PDF Generated successfully: {output_path}")
-            self.logger.info(f"✅ PDF Created: {output_path}")
-        except Exception as e:
-            self.logger.error(f"Failed to generate professional PDF: {e}", exc_info=True)
-            # Fallback to simple image append if needed, but let's trust the new system
-
-    def _determine_pdf_output_path(self):
-        output_path = self.output_path
-        if output_path.exists() and self.get_pdf_url():
-            return self.pdf_dir / f"{self.ms_id}_compiled.pdf"
-        return output_path
-
-    def _load_logo_bytes(self, logo_path):
-        if not logo_path:
-            return None
-        lp = Path(logo_path)
-        if not lp.exists():
-            return None
-        try:
-            return lp.read_bytes()
-        except OSError:
-            self.logger.warning(f"Could not read logo at {logo_path}")
-            return None
-
-    def download_native_pdf(self, pdf_url: str) -> bool:
-        """Download a PDF advertised in the IIIF manifest rendering section."""
-        try:
-            with self.session.get(pdf_url, stream=True, timeout=60) as r:
-                r.raise_for_status()
-                with self.output_path.open("wb") as f:
-                    for chunk in r.iter_content(chunk_size=8192):
-                        if chunk:
-                            f.write(chunk)
-            return True
-        except (RequestException, OSError):
-            self.logger.debug("Native PDF download failed", exc_info=True)
-            return False
-
-    def _should_prefer_native_pdf(self) -> bool:
-        try:
-            return bool(self.cm.get_setting("pdf.prefer_native_pdf", True))
-        except (OSError, ValueError, TypeError):
-            return True
-
-    def _should_create_pdf_from_images(self) -> bool:
-        try:
-            return bool(self.cm.get_setting("pdf.create_pdf_from_images", False))
-        except (OSError, ValueError, TypeError):
-            return False
-
-    def _extract_pages_from_pdf(
-        self,
-        pdf_path: Path,
-        progress_callback: Callable[[int, int], None] | None = None,
-    ) -> bool:
-        self._clear_existing_scans()
-        viewer_dpi = int(self.cm.get_setting("pdf.viewer_dpi", 150) or 150)
-        viewer_quality = int(self.cm.get_setting("images.viewer_quality", 90) or 90)
-        ok, message = convert_pdf_to_images(
-            pdf_path=pdf_path,
-            output_dir=self.scans_dir,
-            progress_callback=progress_callback,
-            dpi=viewer_dpi,
-            jpeg_quality=viewer_quality,
-        )
-        if ok:
-            self.logger.info("Native PDF extraction completed: %s", message)
-            return True
-        self.logger.warning("Native PDF extraction failed: %s", message)
-        self._clear_existing_scans()
-        return False
-
-    def _clear_existing_scans(self) -> None:
-        for scan_file in self.scans_dir.glob("pag_*.jpg"):
-            with suppress(OSError):
-                scan_file.unlink()
-
-    def _collect_scan_stats(self, source_label: str) -> list[dict[str, Any]]:
-        page_stats: list[dict[str, Any]] = []
-        for index, image_path in enumerate(sorted(self.scans_dir.glob("pag_*.jpg"))):
-            try:
-                with Image.open(image_path) as img:
-                    width, height = img.size
-                page_stats.append(
-                    {
-                        "page_index": index,
-                        "filename": image_path.name,
-                        "original_url": source_label,
-                        "thumbnail_url": None,
-                        "size_bytes": image_path.stat().st_size,
-                        "width": width,
-                        "height": height,
-                        "resolution_category": "High" if width > 2500 else "Medium",
-                    }
-                )
-            except Exception:
-                self.logger.debug("Failed to collect scan stats for %s", image_path, exc_info=True)
-        return page_stats
-
-    def _try_native_pdf_flow(self, native_pdf_url: str, progress_callback: Callable[[int, int], None] | None) -> bool:
-        self.logger.info("Manifest advertises native PDF: %s", native_pdf_url)
-        if not self.download_native_pdf(native_pdf_url):
-            self.logger.warning("Native PDF download failed; falling back to canvas image download.")
-            return False
-
-        if not self._extract_pages_from_pdf(self.output_path, progress_callback=progress_callback):
-            self.logger.warning("Native PDF extraction failed; falling back to canvas image download.")
-            return False
-
-        page_stats = self._collect_scan_stats(f"{native_pdf_url} (native-pdf)")
-        self._store_page_stats(page_stats)
-        final_files = [str(path) for path in sorted(self.scans_dir.glob("pag_*.jpg"))]
-        self.vault.upsert_manuscript(
-            self.ms_id,
-            status="complete",
-            asset_state="complete",
-            downloaded_canvases=len(final_files),
-            total_canvases=len(final_files),
-            pdf_local_available=1 if self.output_path.exists() else 0,
-            missing_pages_json="[]",
-            last_sync_at=time.strftime("%Y-%m-%d %H:%M:%S"),
-        )
-        try:
-            clean_dir(self.temp_dir)
-        except OSError:
-            self.logger.debug("Failed to clean temp dir %s", self.temp_dir, exc_info=True)
-
-        if final_files and self.ocr_model:
-            self.run_batch_ocr(final_files, self.ocr_model)
-        return bool(final_files)
-
-    @staticmethod
-    def _build_canvas_plan(canvases: list[dict[str, Any]], target_pages: set[int] | None) -> tuple[set[int], list]:
-        selected_pages = set(target_pages or set())
-        plan: list[tuple[int, dict[str, Any]]] = []
-        for idx, canvas in enumerate(canvases):
-            if selected_pages and (idx + 1) not in selected_pages:
-                continue
-            plan.append((idx, canvas))
-        return selected_pages, plan
-
-    def _mark_downloading_state(self, total_pages: int) -> None:
-        self.vault.upsert_manuscript(
-            self.ms_id,
-            status="downloading",
-            asset_state="downloading",
-            total_canvases=total_pages,
-        )
-
-    def _maybe_run_native_pdf(
-        self,
-        native_pdf_url: str | None,
-        selected_pages: set[int],
-        progress_callback: Callable[[int, int], None] | None,
-    ) -> bool:
-        should_prefer_native_pdf = self._should_prefer_native_pdf()
-        if native_pdf_url and should_prefer_native_pdf and not selected_pages:
-            return self._try_native_pdf_flow(native_pdf_url, progress_callback=progress_callback)
-        if native_pdf_url and not should_prefer_native_pdf:
-            self.logger.info("Native PDF found but preference disabled; proceeding canvas-by-canvas download.")
-        else:
-            self.logger.info("No native PDF rendering found; proceeding canvas-by-canvas download.")
-        return False
-
-    def run(
-        self,
-        progress_callback: Callable[[int, int], None] | None = None,
-        should_cancel: Callable[[], bool] | None = None,
-        target_pages: set[int] | None = None,
-    ):
-        """Execute the download workflow for the manifest.
-
-        Args:
-            progress_callback: Optional callable that receives (current, total)
-                and will be invoked as pages are downloaded. If provided it
-                overrides any instance-level `progress_callback` set at init.
-            should_cancel: Optional callable returning True to request cancellation.
-                It is polled during the run; when True, the method should abort
-                further processing and perform any necessary cleanup.
-            target_pages: Optional 1-based page numbers to retry/download.
-                When omitted, the full manuscript is processed.
-        """
-        self.extract_metadata()
-        canvases = self.get_canvases()
-        total_pages = len(canvases)
-        selected_pages, canvas_plan = self._build_canvas_plan(canvases, target_pages)
-        operation_total = len(canvas_plan)
-        if operation_total == 0:
-            self._sync_asset_state(total_pages)
-            return
-        self._mark_downloading_state(total_pages)
-
-        if self.clean_cache:
-            clean_dir(self.temp_dir)
-
-        # Prefer runtime provided callback over instance attribute
-        cb = progress_callback or self.progress_callback
-
-        if self._maybe_run_native_pdf(self.get_pdf_url(), selected_pages, cb):
-            return
-
-        downloaded, page_stats = self._download_canvases(
-            canvas_plan,
-            progress_callback=cb,
-            should_cancel=should_cancel,
-            total_for_progress=operation_total,
-        )
-        self._store_page_stats(page_stats)
-
-        valid = [f for f in downloaded if f]
-        try:
-            final_files = self._finalize_downloads(valid)
-        except Exception as exc:
-            self.vault.update_status(self.ms_id, "error", str(exc))
-            raise
-
-        if self._should_create_pdf_from_images() and final_files and not selected_pages:
-            self.create_pdf(files=final_files)
-
-        if final_files and self.ocr_model:
-            self.run_batch_ocr(final_files, self.ocr_model)
-        self._sync_asset_state(total_pages)
-
-    def _download_canvases(
-        self,
-        canvas_plan: list[tuple[int, dict[str, Any]]],
-        progress_callback: Callable[[int, int], None] | None = None,
-        should_cancel: Callable[[], bool] | None = None,
-        total_for_progress: int = 0,
-    ):
-        downloaded, page_stats, to_download = self._prescan_canvases(canvas_plan)
-
-        # Initial progress reflecting pre-existing files
-        if progress_callback:
-            try:
-                completed = sum(1 for f in downloaded if f)
-                progress_callback(completed, total_for_progress or len(canvas_plan))
-            except Exception:
-                self.logger.debug("Initial progress callback failed", exc_info=True)
-
-        self.total_canvases = total_for_progress or len(canvas_plan)
-        downloaded, page_stats = self._download_missing_canvases(
-            to_download,
-            downloaded,
-            page_stats,
-            progress_callback,
-            should_cancel,
-            total_for_progress or len(canvas_plan),
-        )
-        return downloaded, page_stats
-
-    def _prescan_canvases(self, canvas_plan: list[tuple[int, dict[str, Any]]]):
-        """Check temp dir for existing valid images and build download list."""
-        downloaded: list[str | None] = [None] * len(canvas_plan)
-        page_stats: list[dict[str, Any]] = []
-        to_download: list[tuple[int, int, dict[str, Any]]] = []
-        for local_idx, (canvas_idx, canvas) in enumerate(canvas_plan):
-            downloader = PageDownloader(self, canvas, canvas_idx, self.temp_dir)
-            resumed = downloader.resume_cached()
-            if resumed:
-                fname, stats = resumed
-                downloaded[local_idx] = fname
-                if stats:
-                    page_stats.append(stats)
-            else:
-                to_download.append((local_idx, canvas_idx, canvas))
-        return downloaded, page_stats, to_download
-
-    def _download_missing_canvases(
-        self,
-        to_download: list[tuple[int, int, dict[str, Any]]],
-        downloaded: list[str | None],
-        page_stats: list[dict[str, Any]],
-        progress_callback: Callable[[int, int], None] | None,
-        should_cancel: Callable[[], bool] | None,
-        total_canvases: int,
-    ):
-        """Download missing pages using ThreadPoolExecutor and update progress and stats."""
-        if not to_download:
-            return downloaded, page_stats
-
-        with ThreadPoolExecutor(max_workers=self.workers) as executor:
-            future_to_index = {
-                executor.submit(self.download_page, canvas, canvas_idx, self.temp_dir): local_idx
-                for local_idx, canvas_idx, canvas in to_download
-            }
-            for future in self._iter_download_futures(future_to_index, len(to_download)):
-                idx = future_to_index[future]
-                try:
-                    result = future.result()
-                except Exception:
-                    result = None
-                if result:
-                    fname, stats = result
-                    downloaded[idx] = fname
-                    if stats:
-                        page_stats.append(stats)
-
-                if progress_callback:
-                    completed = sum(1 for f in downloaded if f)
-                    try:
-                        progress_callback(completed, total_canvases)
-                    except Exception:
-                        self.logger.debug("Progress callback raised an exception", exc_info=True)
-
-                if should_cancel and should_cancel():
-                    completed = sum(1 for f in downloaded if f)
-                    try:
-                        self.vault.update_download_job(
-                            self.job_id or self.ms_id,
-                            current=completed,
-                            total=total_canvases,
-                            status="cancelled",
-                            error="Cancelled by user",
-                        )
-                    except Exception:
-                        self.logger.debug("Failed to mark job cancelled in DB", exc_info=True)
-                    break
-
-        return downloaded, page_stats
-
-    def _iter_download_futures(self, future_to_index: dict, total_to_download: int):
-        futures_iter = as_completed(future_to_index)
-        if self.show_progress:
-            return tqdm(futures_iter, total=total_to_download)
-        return futures_iter
-
-    def _store_page_stats(self, page_stats):
-        if page_stats:
-            page_stats.sort(key=lambda x: x.get("page_index", 0))
-            save_json(self.stats_path, {"doc_id": self.ms_id, "pages": page_stats})
-
-    def _finalize_downloads(self, valid):
-        final_files = []
-        for temp_file in valid:
-            p = Path(temp_file)
-            dest = self.scans_dir / p.name
-            if not dest.exists():
-                shutil.move(str(p), str(dest))
-            elif self.overwrite_existing_scans:
-                shutil.copy2(str(p), str(dest))
-                with suppress(OSError):
-                    p.unlink()
-            else:
-                # In case of partial resume or overwrite
-                pass
-            final_files.append(str(dest))
-        # Clean up the temporary directory that held the downloaded images.
-        # This removes the folder that contained the images (and any remaining files).
-        try:
-            clean_dir(self.temp_dir)
-        except OSError:
-            # Avoid raising during cleanup; log and continue.
-            with suppress(Exception):
-                self.logger.debug("Failed to clean temp dir %s", self.temp_dir, exc_info=True)
-
-        return final_files
-
-    def _sync_asset_state(self, total_expected: int) -> None:
-        scan_count = len(list(self.scans_dir.glob("pag_*.jpg")))
-        pdf_available = 1 if any(self.pdf_dir.glob("*.pdf")) else 0
-        if scan_count <= 0 and total_expected > 0:
-            state = "saved"
-        elif total_expected <= 0 or scan_count >= total_expected:
-            state = "complete"
-        else:
-            state = "partial"
-        missing = []
-        if total_expected > 0 and scan_count < total_expected:
-            existing = {int(p.stem.split("_")[-1]) + 1 for p in self.scans_dir.glob("pag_*.jpg")}
-            missing = [i for i in range(1, total_expected + 1) if i not in existing]
-        self.vault.upsert_manuscript(
-            self.ms_id,
-            status=state,
-            asset_state=state,
-            total_canvases=total_expected,
-            downloaded_canvases=scan_count,
-            pdf_local_available=pdf_available,
-            missing_pages_json=json.dumps(missing),
-            last_sync_at=time.strftime("%Y-%m-%d %H:%M:%S"),
-        )
-
-    def run_batch_ocr(self, image_files: list[str], model_name: str):
-        """Run OCR processing on the finalized image files."""
-        from ..services.ocr.model_manager import ModelManager
-        from ..services.ocr.processor import KRAKEN_AVAILABLE, OCRProcessor
-
-        if not KRAKEN_AVAILABLE:
-            return
-        manager = ModelManager()
-        model_path = manager.get_model_path(model_name)
-        proc = OCRProcessor(model_path)
-        aggregated = {
-            "metadata": {
-                "manuscript_id": self.ms_id,
-                "model": self.ocr_model,
-                "processed_at": time.strftime("%Y-%m-%d %H:%M:%S"),
-            },
-            "pages": [],
-        }
-        for i, img_path in enumerate(image_files):
-            try:
-                res = proc.process_image(img_path)
-                if "error" not in res:
-                    res["page_index"] = i + 1
-                    aggregated["pages"].append(res)
-            except Exception:
-                self.logger.debug("OCR failed for %s", img_path, exc_info=True)
-        save_json(self.ocr_path, aggregated)
+__all__ = ["CanvasServiceLocator", "PageDownloader", "IIIFDownloader"]
