+++ src/universal_iiif_core/services/storage/vault_manager.py
@@ -5,8 +5,6 @@ from contextlib import suppress
 from pathlib import Path
 from typing import Any
 
-import fitz  # PyMuPDF
-
 from ...exceptions import DatabaseError
 from ...library_catalog import normalize_item_type
 from ...logger import get_logger
@@ -698,660 +696,11 @@ class VaultManager:
         finally:
             conn.close()
 
-    def extract_image_snippet(self, image_path: str, coordinates: tuple) -> bytes | None:
-        """Extracts a crop from an image using PyMuPDF (fitz).
 
-        coordinates: (x0, y0, x1, y1) relative to the original image size.
-        """
-        try:
-            # We assume image_path is a valid image file.
-            # PyMuPDF can open images as simple documents.
-            with fitz.open(image_path) as doc:
-                page = doc.load_page(0)  # Images are 1-page documents
-                rect = fitz.Rect(coordinates)
+from .vault_jobs import attach_job_methods  # noqa: E402
+from .vault_snippets import attach_snippet_methods  # noqa: E402
 
-                # Pixmap of the cropped area
-                pix = page.get_pixmap(clip=rect)
+attach_snippet_methods(VaultManager)
+attach_job_methods(VaultManager)
 
-                # Convert to PNG bytes
-                return pix.tobytes("png")
-        except Exception as e:
-            logger.error(f"Error extracting snippet with fitz: {e}", exc_info=True)
-            return None
-
-    def save_snippet(
-        self,
-        ms_name: str,
-        page_num: int,
-        image_path: str,
-        category: str = "Uncategorized",
-        transcription: str | None = None,
-        notes: str | None = None,
-        coords: list[Any] | None = None,
-    ):
-        """Saves the snippet to the database.
-
-        Args:
-            ms_name: Nome del manoscritto
-            page_num: Numero pagina (1-indexed)
-            image_path: Path del file immagine salvato
-            category: Categoria (Capolettera, Glossa, etc)
-            transcription: Trascrizione rapida
-            notes: Note/commenti
-            coords: Coordinate [x, y, width, height]
-        """
-        import json
-
-        logger.debug(f"💾 SAVE_SNIPPET - ms_name='{ms_name}', page_num={page_num}, category='{category}'")
-
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        try:
-            coords_json = json.dumps(coords) if coords else None
-            cursor.execute(
-                """
-                INSERT INTO snippets (ms_name, page_num, image_path, category, transcription, notes, coords_json)
-                VALUES (?, ?, ?, ?, ?, ?, ?)
-            """,
-                (ms_name, page_num, image_path, category, transcription, notes, coords_json),
-            )
-            conn.commit()
-            snippet_id = cursor.lastrowid
-            logger.info(f"✅ Snippet ID {snippet_id} salvato: {category} - p.{page_num}")
-
-            return snippet_id
-        finally:
-            conn.close()
-
-    def get_snippets(self, ms_name: str, page_num: int | None = None):
-        """Retrieve snippets for a manuscript, optionally filtered by page.
-
-        Args:
-            ms_name: Nome del manoscritto
-            page_num: Numero pagina opzionale (1-indexed)
-
-        Returns:
-            List of snippet dictionaries
-        """
-        import json
-
-        logger.debug(f"🔍 GET_SNIPPETS - ms_name='{ms_name}', page_num={page_num}")
-
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        try:
-            if page_num is not None:
-                query = """
-                    SELECT id, ms_name, page_num, image_path, category, transcription,
-                           notes, coords_json, timestamp
-                    FROM snippets
-                    WHERE ms_name = ? AND page_num = ?
-                    ORDER BY timestamp DESC
-                """
-                params = (ms_name, page_num)
-                cursor.execute(query, params)
-            else:
-                query = """
-                    SELECT id, ms_name, page_num, image_path, category, transcription,
-                           notes, coords_json, timestamp
-                    FROM snippets
-                    WHERE ms_name = ?
-                    ORDER BY page_num, timestamp DESC
-                """
-                params = (ms_name,)
-                cursor.execute(query, params)
-
-            rows = cursor.fetchall()
-            logger.debug(f"Query eseguita: trovati {len(rows)} snippet")
-            snippets = []
-            for row in rows:
-                snippets.append(
-                    {
-                        "id": row[0],
-                        "ms_name": row[1],
-                        "page_num": row[2],
-                        "image_path": row[3],
-                        "category": row[4],
-                        "transcription": row[5],
-                        "notes": row[6],
-                        "coords_json": json.loads(row[7]) if row[7] else None,
-                        "timestamp": row[8],
-                    }
-                )
-            return snippets
-        finally:
-            conn.close()
-
-    def delete_snippet(self, snippet_id: int):
-        """Delete a snippet by ID and remove the physical file."""
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        try:
-            # Get image path before deleting
-            cursor.execute("SELECT image_path FROM snippets WHERE id = ?", (snippet_id,))
-            result = cursor.fetchone()
-
-            if result and result[0]:
-                image_path = Path(result[0])
-                if image_path.exists():
-                    with suppress(OSError):
-                        image_path.unlink()
-
-            # Delete from DB
-            cursor.execute("DELETE FROM snippets WHERE id = ?", (snippet_id,))
-            conn.commit()
-        finally:
-            conn.close()
-
-    def create_download_job(self, job_id: str, doc_id: str, library: str, manifest_url: str):
-        """Crea traccia del download nel DB."""
-        conn = self._get_conn()
-        try:
-            cursor = conn.cursor()
-            # Creiamo la tabella al volo se non esiste (safety check)
-            cursor.execute("""
-                CREATE TABLE IF NOT EXISTS download_jobs (
-                    job_id TEXT PRIMARY KEY, doc_id TEXT, library TEXT,
-                    manifest_url TEXT, status TEXT, current_page INTEGER,
-                    total_pages INTEGER, queue_position INTEGER DEFAULT 0,
-                    priority INTEGER DEFAULT 0, error_message TEXT, started_at TIMESTAMP,
-                    finished_at TIMESTAMP, updated_at TIMESTAMP
-                )
-            """)
-            cursor.execute(
-                """
-                INSERT OR REPLACE INTO download_jobs (
-                    job_id, doc_id, library, manifest_url, status,
-                    current_page, total_pages, queue_position, priority, started_at, finished_at, updated_at
-                )
-                VALUES (?, ?, ?, ?, 'queued', 0, 0, 0, 0, NULL, NULL, CURRENT_TIMESTAMP)
-            """,
-                (job_id, doc_id, library, manifest_url),
-            )
-            conn.commit()
-        finally:
-            conn.close()
-
-    def update_download_job(
-        self,
-        job_id: str,
-        current: int,
-        total: int,
-        status: str = "running",
-        error: str | None = None,
-        queue_position: int | None = None,
-        priority: int | None = None,
-    ):
-        """Aggiorna lo stato."""
-        conn = self._get_conn()
-        try:
-            cursor = conn.cursor()
-            updates = [
-                "current_page = ?",
-                "total_pages = ?",
-                "status = ?",
-                "error_message = ?",
-                "updated_at = CURRENT_TIMESTAMP",
-            ]
-            params: list[Any] = [current, total, status, error]
-            if queue_position is not None:
-                updates.append("queue_position = ?")
-                params.append(int(queue_position))
-            if priority is not None:
-                updates.append("priority = ?")
-                params.append(int(priority))
-            if status == "running":
-                updates.append("started_at = COALESCE(started_at, CURRENT_TIMESTAMP)")
-                updates.append("finished_at = NULL")
-            if status in {"completed", "error", "cancelled", "paused"}:
-                updates.append("finished_at = CURRENT_TIMESTAMP")
-            sql = f"UPDATE download_jobs SET {', '.join(updates)} WHERE job_id = ?"  # noqa: S608
-            params.append(job_id)
-            cursor.execute(sql, tuple(params))
-            progress_key = (current, total, status)
-            cached = self._download_progress_cache.get(job_id)
-            if cached != progress_key:
-                try:
-                    cursor.execute("SELECT doc_id FROM download_jobs WHERE job_id = ?", (job_id,))
-                    row = cursor.fetchone()
-                    doc_id = row[0] if row else None
-                    title = None
-                    if doc_id:
-                        try:
-                            ms = self.get_manuscript(doc_id)
-                            title = ms.get("title") if ms else None
-                        except DatabaseError:
-                            title = None
-
-                    log_message = "Download update: job=%s doc=%s title=%s %s/%s status=%s"
-                    log_args = (job_id, doc_id or "-", title or "-", current, total, status)
-                    if status in {"completed", "error", "cancelled", "cancelling", "paused"}:
-                        logger.info(log_message, *log_args)
-                    else:
-                        logger.debug(log_message, *log_args)
-                except DatabaseError:
-                    logger.debug("Failed to log download update for job %s", job_id, exc_info=True)
-                finally:
-                    self._download_progress_cache[job_id] = progress_key
-            conn.commit()
-        finally:
-            conn.close()
-
-    def get_active_download(self):
-        """Trova un download attivo."""
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        # Controlla se la tabella esiste prima di fare select
-        cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='download_jobs'")
-        if not cursor.fetchone():
-            return None
-
-        cursor.execute("""
-            SELECT job_id, doc_id, library, status, current_page, total_pages, error_message
-            FROM download_jobs
-            WHERE status IN ('running', 'queued')
-            ORDER BY priority DESC, queue_position ASC, updated_at DESC LIMIT 1
-        """)
-        row = cursor.fetchone()
-        if row:
-            return {
-                "job_id": row[0],
-                "doc_id": row[1],
-                "library": row[2],
-                "status": row[3],
-                "current": row[4],
-                "total": row[5],
-                "error": row[6],
-            }
-        return None
-
-    def get_active_downloads(self):
-        """Return a list of active (pending/running) download jobs ordered by recent update.
-
-        This is used by the UI to render all concurrent downloads.
-        """
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='download_jobs'")
-        if not cursor.fetchone():
-            return []
-
-        cursor.execute(
-            """
-            SELECT job_id, doc_id, library, status, current_page, total_pages, error_message, queue_position, priority
-            FROM download_jobs
-            WHERE status IN ('running', 'queued', 'cancelling')
-            ORDER BY CASE
-                WHEN status = 'running' THEN 0
-                WHEN status = 'cancelling' THEN 1
-                ELSE 2
-            END, priority DESC, queue_position ASC, updated_at DESC
-        """
-        )
-        rows = cursor.fetchall()
-        conn.close()
-        results = []
-        for row in rows:
-            results.append(
-                {
-                    "job_id": row[0],
-                    "doc_id": row[1],
-                    "library": row[2],
-                    "status": row[3],
-                    "current": row[4],
-                    "total": row[5],
-                    "error": row[6],
-                    "queue_position": row[7],
-                    "priority": row[8],
-                }
-            )
-        return results
-
-    def get_download_job(self, job_id: str):
-        """Return a specific download job by its job_id, or None if not found."""
-        conn = self._get_conn()
-        conn.row_factory = None
-        cursor = conn.cursor()
-        cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='download_jobs'")
-        if not cursor.fetchone():
-            return None
-        cursor.execute(
-            """
-            SELECT job_id, doc_id, library, manifest_url, status, current_page, total_pages,
-                   error_message, queue_position, priority
-            FROM download_jobs
-            WHERE job_id = ?
-        """,
-            (job_id,),
-        )
-        row = cursor.fetchone()
-        conn.close()
-        if not row:
-            return None
-        return {
-            "job_id": row[0],
-            "doc_id": row[1],
-            "library": row[2],
-            "manifest_url": row[3],
-            "status": row[4],
-            "current": row[5],
-            "total": row[6],
-            "error": row[7],
-            "queue_position": row[8],
-            "priority": row[9],
-        }
-
-    def list_download_jobs(self, limit: int = 50):
-        """List recent download jobs with queue metadata."""
-        conn = self._get_conn()
-        conn.row_factory = sqlite3.Row
-        cursor = conn.cursor()
-        try:
-            cursor.execute(
-                """
-                SELECT dj.job_id, dj.doc_id, dj.library, dj.manifest_url, dj.status,
-                       current_page AS current, total_pages AS total,
-                       dj.queue_position, dj.priority, dj.error_message AS error,
-                       dj.updated_at, dj.created_at, dj.started_at, dj.finished_at,
-                       m.display_title, m.catalog_title, m.shelfmark
-                FROM download_jobs dj
-                LEFT JOIN manuscripts m
-                    ON m.id = dj.doc_id AND COALESCE(m.library, '') = COALESCE(dj.library, '')
-                WHERE NOT (
-                    dj.status IN ('completed', 'error', 'cancelled', 'paused')
-                    AND m.id IS NULL
-                )
-                ORDER BY
-                    CASE
-                        WHEN dj.status = 'running' THEN 0
-                        WHEN dj.status = 'queued' THEN 1
-                        WHEN dj.status = 'cancelling' THEN 2
-                        WHEN dj.status = 'paused' THEN 3
-                        WHEN dj.status = 'error' THEN 4
-                        ELSE 5
-                    END,
-                    dj.priority DESC,
-                    dj.queue_position ASC,
-                    dj.updated_at DESC
-                LIMIT ?
-                """,
-                (int(limit),),
-            )
-            return [dict(row) for row in cursor.fetchall()]
-        finally:
-            conn.close()
-
-    def delete_download_job(self, job_id: str) -> bool:
-        """Delete one download job row by id."""
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        try:
-            cursor.execute("DELETE FROM download_jobs WHERE job_id = ?", (job_id,))
-            deleted = cursor.rowcount > 0
-            conn.commit()
-            return deleted
-        finally:
-            conn.close()
-
-    def create_export_job(
-        self,
-        job_id: str,
-        *,
-        scope_type: str,
-        doc_ids_json: str,
-        library: str,
-        export_format: str,
-        output_kind: str,
-        selection_mode: str,
-        selected_pages_json: str,
-        destination: str,
-        destination_payload_json: str = "{}",
-        capability_flags_json: str = "{}",
-        total_steps: int = 0,
-    ) -> None:
-        """Create or replace one export job row."""
-        conn = self._get_conn()
-        try:
-            cursor = conn.cursor()
-            cursor.execute(
-                """
-                INSERT OR REPLACE INTO export_jobs (
-                    job_id, scope_type, doc_ids_json, library, export_format, output_kind,
-                    selection_mode, selected_pages_json, destination, destination_payload_json,
-                    capability_flags_json, status, current_step, total_steps, output_path, error_message,
-                    started_at, finished_at, created_at, updated_at
-                )
-                VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 'queued', 0, ?, NULL, NULL, NULL, NULL,
-                        CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)
-                """,
-                (
-                    job_id,
-                    scope_type,
-                    doc_ids_json,
-                    library,
-                    export_format,
-                    output_kind,
-                    selection_mode,
-                    selected_pages_json,
-                    destination,
-                    destination_payload_json,
-                    capability_flags_json,
-                    int(total_steps),
-                ),
-            )
-            conn.commit()
-        finally:
-            conn.close()
-
-    def update_export_job(
-        self,
-        job_id: str,
-        *,
-        current_step: int | None = None,
-        total_steps: int | None = None,
-        status: str | None = None,
-        output_path: str | None = None,
-        error_message: str | None = None,
-    ) -> None:
-        """Update one export job row with optional fields."""
-        updates: list[str] = ["updated_at = CURRENT_TIMESTAMP"]
-        params: list[Any] = []
-
-        if current_step is not None:
-            updates.append("current_step = ?")
-            params.append(int(current_step))
-        if total_steps is not None:
-            updates.append("total_steps = ?")
-            params.append(int(total_steps))
-        if status is not None:
-            updates.append("status = ?")
-            params.append(status)
-            if status == "running":
-                updates.append("started_at = COALESCE(started_at, CURRENT_TIMESTAMP)")
-                updates.append("finished_at = NULL")
-            if status in {"completed", "error", "cancelled"}:
-                updates.append("finished_at = CURRENT_TIMESTAMP")
-        if output_path is not None:
-            updates.append("output_path = ?")
-            params.append(output_path)
-        if error_message is not None:
-            updates.append("error_message = ?")
-            params.append(error_message)
-
-        conn = self._get_conn()
-        try:
-            cursor = conn.cursor()
-            sql = f"UPDATE export_jobs SET {', '.join(updates)} WHERE job_id = ?"  # noqa: S608
-            params.append(job_id)
-            cursor.execute(sql, tuple(params))
-            conn.commit()
-        finally:
-            conn.close()
-
-    def get_export_job(self, job_id: str) -> dict[str, Any] | None:
-        """Return one export job by id."""
-        conn = self._get_conn()
-        conn.row_factory = sqlite3.Row
-        cursor = conn.cursor()
-        try:
-            cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='export_jobs'")
-            if not cursor.fetchone():
-                return None
-            cursor.execute("SELECT * FROM export_jobs WHERE job_id = ?", (job_id,))
-            row = cursor.fetchone()
-            return dict(row) if row else None
-        finally:
-            conn.close()
-
-    def list_export_jobs(self, limit: int = 50) -> list[dict[str, Any]]:
-        """List recent export jobs ordered by status and recency."""
-        conn = self._get_conn()
-        conn.row_factory = sqlite3.Row
-        cursor = conn.cursor()
-        try:
-            cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='export_jobs'")
-            if not cursor.fetchone():
-                return []
-            cursor.execute(
-                """
-                SELECT *
-                FROM export_jobs
-                ORDER BY
-                    CASE
-                        WHEN status = 'running' THEN 0
-                        WHEN status = 'queued' THEN 1
-                        WHEN status = 'completed' THEN 2
-                        WHEN status = 'error' THEN 3
-                        WHEN status = 'cancelled' THEN 4
-                        ELSE 5
-                    END,
-                    updated_at DESC
-                LIMIT ?
-                """,
-                (int(limit),),
-            )
-            return [dict(row) for row in cursor.fetchall()]
-        finally:
-            conn.close()
-
-    def delete_export_job(self, job_id: str) -> bool:
-        """Delete one export job row by id."""
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        try:
-            cursor.execute("DELETE FROM export_jobs WHERE job_id = ?", (job_id,))
-            deleted = cursor.rowcount > 0
-            conn.commit()
-            return deleted
-        finally:
-            conn.close()
-
-    def reset_active_downloads(self, mark: str = "error", message: str = "Server restarted"):
-        """Mark any downloads left in 'pending' or 'running' state as stopped.
-
-        This is intended to be called at application startup so that stale
-        download jobs don't cause the UI to continue polling indefinitely.
-        """
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        try:
-            cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='download_jobs'")
-            if not cursor.fetchone():
-                return 0
-
-            cursor.execute(
-                """
-                UPDATE download_jobs
-                SET status = ?,
-                    error_message = COALESCE(error_message, ?) || ' (server restart)',
-                    updated_at = CURRENT_TIMESTAMP
-                WHERE status IN ('queued', 'pending', 'running')
-            """,
-                (mark, message),
-            )
-            count = cursor.rowcount
-            conn.commit()
-            return count
-        finally:
-            conn.close()
-
-    def reset_active_exports(self, mark: str = "error", message: str = "Server restarted") -> int:
-        """Mark any exports left in queued/running state as stopped on startup."""
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        try:
-            cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='export_jobs'")
-            if not cursor.fetchone():
-                return 0
-
-            cursor.execute(
-                """
-                UPDATE export_jobs
-                SET status = ?,
-                    error_message = COALESCE(error_message, ?) || ' (server restart)',
-                    finished_at = CURRENT_TIMESTAMP,
-                    updated_at = CURRENT_TIMESTAMP
-                WHERE status IN ('queued', 'running')
-                """,
-                (mark, message),
-            )
-            count = cursor.rowcount
-            conn.commit()
-            return count
-        finally:
-            conn.close()
-
-    def cleanup_stale_data(self, retention_hours: int = 24) -> int:
-        """Remove download job records older than `retention_hours` and prune their temp folders.
-
-        Returns the number of jobs removed.
-        """
-        from ...config_manager import get_config_manager
-
-        conn = self._get_conn()
-        cursor = conn.cursor()
-        try:
-            cursor.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='download_jobs'")
-            if not cursor.fetchone():
-                return 0
-
-            cursor.execute(
-                "SELECT job_id, doc_id FROM download_jobs WHERE created_at < datetime('now', ?)",
-                (f"-{int(retention_hours)} hours",),
-            )
-            rows = cursor.fetchall()
-            if not rows:
-                return 0
-
-            job_ids = [r[0] for r in rows]
-            doc_ids = [r[1] for r in rows if r[1]]
-
-            # Delete DB records
-            placeholders = ",".join("?" for _ in job_ids)
-            sql = f"DELETE FROM download_jobs WHERE job_id IN ({placeholders})"  # noqa: S608
-            cursor.execute(sql, tuple(job_ids))
-            conn.commit()
-
-            # Prune temp dirs for associated doc_ids
-            try:
-                cm = get_config_manager()
-                base_temp = cm.get_temp_dir()
-                for did in set(doc_ids):
-                    if not did:
-                        continue
-                    p = base_temp / str(did)
-                    if p.exists() and p.is_dir():
-                        try:
-                            # Use shutil.rmtree to remove contents
-                            import shutil
-
-                            shutil.rmtree(p)
-                            logger.info("Pruned stale temp folder: %s", p)
-                        except OSError:
-                            logger.debug("Failed to remove temp folder %s", p, exc_info=True)
-            except OSError:
-                logger.debug("Failed to cleanup temp dirs for stale jobs", exc_info=True)
-
-            return len(job_ids)
-        finally:
-            conn.close()
+__all__ = ["VaultManager"]
